requirements.txt
torch>=2.0
opencv-python>=4.8
pillow>=10.0
numpy>=1.24
python-dotenv>=1.0
vision-agents[getstream,gemini]>=0.3.8
ultralytics>=8.0
google-generativeai>=0.10.0
gTTS>=2.4.0
pyttsx3>=2.90
pydantic>=2.0